# Add verbose information (config/system/performance)