Support displaying verbose output including config, system and performance information. Enable it by passing `-v` or `--verbose` to the `main` program. This should resolve #8.

Sample output:
```
$ ./build/bin/main -m chatglm2-ggml.bin -p 你好 -v -t 16
system info: | AVX = 1 | AVX2 = 1 | AVX512 = 0 | AVX512_VBMI = 0 | AVX512_VNNI = 0 | FMA = 1 | NEON = 0 | ARM_FMA = 0 | F16C = 1 | FP16_VA = 0 | WASM_SIMD = 0 | BLAS = 0 | SSE3 = 1 | VSX = 0 |
inference config: | max_length = 2048 | max_context_length = 512 | top_k = 0 | top_p = 0.7 | temperature = 0.95 | num_threads = 16 |
loaded ChatGLM2 model from chatglm2-ggml.bin within: 25.119 ms

你好👋！我是人工智能助手 ChatGLM2-6B，很高兴见到你，欢迎问我任何问题。

prompt time: 633.128 ms / 17 tokens (37.242 ms/token)
output time: 2202.78 ms / 28 tokens (78.67 ms/token)
total time: 2835.91 ms
```